Cleanup: Rename some views, remove SDWebImageSwiftUI and create new component.
